refactor(blorktools): html-editor-modal directory
