+++ packages/blorktools/src/asset_debugger/core/animation/threejs-util.js
@@ -9,7 +9,7 @@ import {
 } from './animation-util';
 import { createTextureFromIframe } from '../texture-util';
 import { getState } from '../state';
-import { showStatus } from '../../ui/scripts/html-editor-modal';
+import { showStatus } from '../../ui/html-editor-modal/html-editor-modal';
 import { createMeshInfoPanel, infoPanel, resetInfoPanel } from '../mesh-info-panel-util';
 
 let pendingTextureUpdate = false;
